Fix ShellCommandPermissionTest
The `ShellCommandPermission` tests previously checked `stderr` for a
line containing the string `Error:` to determine if the command failed
to run or not. Due to recent changes no output is generated when
checking whether the command can run, instead the process returns an
exit code of -1. This updates the permissions tests to check that the
process exits with an error code instead of relying on parsing the
output of stderr.

Bug: 288217526
Bug: 303787380
Change-Id: I2a5b1d4ac19d8def11a7775f56c23d8befb433b7
Fixes: 293898385
Fixes: 289277264
Fixes: 293799054
Test: atest CtsPermissionTestCases:android.permission.cts.ShellCommandPermissionTest
Merged-In: If8856b546a8d4a733e3ec0139798b4495e604054