Deflake MediaControllerTest#testOnBackInvokedCallback.
The root cause of the flake: There can be a delay between
MediaController requesting to show itself and its window getting focus.
When KEYCODE_BACK is dispatched before MediaController's window is in
focus, the back key is sent to the activity instead. When this happens,
the activity is dismissed instead of MediaController being closed.

The fix is to wait for the activity window to lose focus before sending
KEYCODE_BACK.

Bug: 301736426
Test: atest MediaControllerTest#testOnBackInvokedCallback --rerun-until-failure 100
Test: atest MediaControllerTest
Change-Id: Ib2cfe102a940ef3c99c57e6eeed9c2ddb3db6e7b