Wait for color changes in WallpaperManagerTest#setUp
This prevents some race condition on cuttlefish: the tests for color
listeners would fail about 4-5% of times, because the color events from
WallpaperManager#clear (done during setup) would be called during the
test and fail the test to an unwanted callback.

This is mainly a copy of ag/24803244 (patchset 1). Kudos to erahm@ for
writing easy-to-steal code :)

Bug: 301734566
Test: atest WallpaperManagerTest --iteration 20 (on cuttleflish)
Test: atest WallpaperManagerTest#invokeOnColorsChangedListenerTest_systemOnly WallpaperManagerTest#invokeOnColorsChangedListenerTest_lockOnly --iterations 500 (on cuttlefish)

Change-Id: I5428768912587cb8d0a406bf4502316e15a2f01e