Further checking of external storage status
Some processes seemed not to go through the polling loop and would
get an invalid storage state. Add code that will detect when the storage
state has reverted after being marked mounted and re-run the polling
loop. Again, all of this to handle what appears to be much slower
performance while doing cuttlefish/coverage runs; this is the only place
that we see any of this behavior.

Bug: 305448593
Test: CtsMediaMiscTestCases
Change-Id: I99bb72d2f089b0c5b5e35e009c4006bb9ebbf356