Improve test invokeOnColorsChangedListenerTest_clearLock
Currently, the test sets a shared bitmap on home & lock screen, then
clears the wallpaper on either home, lock, or home+lock, then checks
that we receive color events for the clear operation.

This is dubious for the clear(lock) case: clear(lock) means "set the
lock screen wallpaper to be the same as the home screen wallpaper".
Since we set a shared bitmap before clearing, the lock screen wallpaper
is already the same and this clear(lock) is a no-op. It is absolutely
fine not to receive a color event in that case.

This CLs sets a bitmap only to the destination that is cleared, before
running the checks on the clear method.

Flag: NONE
Test: atest WallpaperManagerTest --iterations 20
Bug: 308757171
Change-Id: I8ff8ce28f6d3b18d8326b111938fbedc35d3e0bb