Align time reporting tolerance in AudioTrack tests
There are multiple tests that assume certain tolerance
for playback position reporting:

 - AudioTrackTest#testVariableRatePlayback: 60 ms;
 - AudioTrackTest#testVariableSpeedPlayback: 100 ms;
 - AudioTrack_ListenerTest: 80 ms.

Update all to match the highest figure when the "low latency"
audio feature is not supported. This allows these tests
to pass on CF with AIDL HAL. The test was failing because the
tolerance value ended up being smaller than the HAL buffer size.

Bug: 302132812
Bug: 311745649
Test: atest CtsMediaAudioTestCases --test-filter=".*AudioTrackTest#testVariable.*"
Test: atest CtsMediaAudioTestCases --test-filter=".*AudioTrack_ListenerTest"
Change-Id: Ieced3d4a0f7ae1d72c34940f43f8a7aa8b020387