[Exo] Correct the area for gesture detection
Test testTriggerTouchExploration_topFocusDisplayIsLastNonProxyDisplay would always return false when making an assumption check on the width of the display and would never actually run.

The virtual display size was not the real phone display size, which the
original min width value was based on (taken from TouchExplorerTest).

Since all we're doing is testing taps, which don't require a big area,
we'll just use half of the virtual display size.

Bug: 286917088
Test: AccessibilityDisplayProxyTest
Change-Id: I57a2c6a329a2f9304df91baa2eaac83a0363b7b6