Don't use runtime permission op code for collection test
The appop modes for those associated with runtime permissions default to
MODE_ALLOWED. In this test OP_WIFI_SCAN was used which switches to
OP_COARSE_LOCATION which is associated with the coarse location
permission. This means when the appop is checked and the package doesn't
request that permission then the result will be MODE_ALLOWED. If we want
to switch to using runtime permissions to compute the appop mode for
these at the moment of doing the check then we will see this behavior
change since not requesting means the permission will never be granted.

This test picks another package already on the device to do the appop
blaming. Testing found it was blaming the package installer which
doesn't request coarse location so the checks were successful and
blaming went through.

The solution for this test is to switch to using an appop that isn't
associated with a runtime permission but rather an appop permission so
noteOps get collected, has default mode as MODE_ALLOWED, and is test
API; OPSTR_ACCESS_NOTIFICATIONS seems to be the only one that qualifies.
Possible future flakiness could be prevented if the test owned the other
package instead of resolving a preinstalled package since that package
could always change in ways we don't own.

Test: AppOpEventCollectionTest
Bug: 266164193
Change-Id: Ie42e58dfe722902970309e63bbc3e4894e91a031