DO NOT MERGE: Fix the number of clicks in WindowInsetsBehaviorTests
Before this CL, the number of clicks depended on the display size. The
test clicked more on larger displays, and took more time to complete.
However, the test would time out after 5 minutes. If the test runs on a
large enough display, it would time out.

This CL fixes the number of clicks. So the test time won't be affected
by the display size.

Bug: 276401160
Bug: 311303545
Test: Run WindowInsetsBehaviorTests on a large-display device.
Change-Id: I92494208cae52c8e075e6c1b1e852bfa3cda447a
Signed-off-by: AOYAGI Takako <aoyagi.takako@jp.panasonic.com>