Reduce threshold for GYROSCOPE_UNCALIBRATED.
Reducing the minimum mean value in batch tests for
GYROSCOPE_UNCALIBRATED from 0.0005 rad/s to 0.0002 rad/s. It was found
that a well performing thermal compensation would cause this test to
become flaky. This reduces the threshold to allow such devices to
reliably pass the test while still allow the test to catch egregious
failures.

Bug: 296489011
Test: CTS SensorBatchingTests
Change-Id: Iab512b6cb918c92dca32a5f3c1cae41cbaef3154