Media: Skip few color conversion related tests in MTS
Prior to Android U, tests in DecodeEditEncodeTest and EncodeDecodeTest
only checked the 1st codec (which is usually a hardware codec) and
software codecs exercised a problem in the underlying graphis code.
So we will only run this for CTS mode or if we're on versions after
Android T (where the graphics code is fixed).

Bug: 289303019
Test: atest CtsMediaCodecTestCases:DecodeEditEncodeTest \
 CtsMediaCodecTestCases:EncodeDecodeTest -- --module-arg \
 CtsMediaCodecTestCases:instrumentation-arg:mts-media:=true

Change-Id: I62b43b380b94ddb8aba29b683839df988fd83caf