Modift IntentTest to make it permissive
Logic change in the PackageInstaller app caused a change in CTS test behaviour in midst of a release. This logic change was delivered to select partners via a prebuilt, thus there was an inconsistency in when the test passes / fails.
This change attempts at relieving this inconsistency to an extent.

Bug: 308866556
Change-Id: I9da5423d631f782c5fe14b1e520db95d200ac4a9
Test: atest CtsPackageInstallTestCases:IntentTest