Fix MediaProjection CTS tests for initial platform release
CTS tests started failing due to a change where Partial Screenshare was set as the default option for MediaProjection (ag/24887989), and then an accompanying test fix which looked for the screen share option spinner (ag/25185979). CTS tests still need to work on the intial platform release, which has PSS was disabled and no spinner at all. CTS were failing due to not being able to find the spinner, so this change allows the test to not fail if the spinner does not exist (due to it running on udc-release).

Test: atest CtsMediaProjectionTestCases
Test: atest CtsMediaAudioTestCases
Flag: None
Bug: incoming
Change-Id: Icdbfc34c0f60899b9c667fbfc26898682aa6d77f