Grant BAL privileges when starting for result
When an app starts an Activity for a result (e.g. using
Activity.startIntentSenderForResult) and passes in a valid (>=0)
resultCode, we consider this as an implicit opt-in to start that
Activity from the background using the the sender's privileges.

This changes existing tests to not use `sendIntentSenderForResult`
methods and adds a test to check that these are actually granting an
automatic opt in.

Test: atest BackgroundActivityLaunchTest
Bug: 314033453
Change-Id: If580cf736b4ac65055adf963f6a76fa58b9fa4b0