Add a test to check deletion of app in the clone user
When deleteAppWithParent property is true, the app in the clone user
needs to be removed when the same app in the parent user is deleted.
However, it should not be deleted when the same app in the other user
which has same parent is deleted.

Bug: 308362817
Test: atest com.android.cts.appcloning.AppCloningHostTest
(cherry picked from https://partner-android-review.googlesource.com/q/commit:900b80c0ce827b3da1262bb9bd949dee8c1bf76b)
Change-Id: I4c6c7acb8ec81a83d1117e0675a6b48710940c4e