Remove CtsSystemApiAnnotationTestCases from cts
The test is now removed from CTS because it cannot be correctly
implemented under Trunk stable. The test was designed with the
assumption that CTS and DUT are built from the same branch, or from the
branches having identical set of APIs (like udc-release and
android14-tests-dev).

However, the assumption no longer holds with Trunk stable. As of now,
DUT is built from main, while CTS is from android14-tests-dev. The main
branch has many APIs added for the next release.

In theory, those APIs should be guarded using the FlaggedApi annotation
and the test should not include them when gathering APIs from DUT.
However, this is impossible at the moment for the following two reasons:

1. If a System API is promoted to a public API, it can't be annotated
   with FlaggedApi. That's a known limitation.

2. A newly added SystemApi are properly guarded with FlaggedApi but the
   annotation is not accessible at runtime. The retention policy is set
to SOURCE, which makes it immpossible for the test to filter those out.

Therefore, we are (hopefully temporarily) removing the test from CTS
until the underlying issue is fixed.

Bug: 307877675
Test: cts-tradefed run commitAndExit cts -m CtsSystemApiAnnotationTestCases

======= Summary ===============
Total Run time: 26s
0/0 modules completed
Total Tests       : 0
PASSED            : 0
FAILED            : 0
============== End of Results ==============
============================================

Change-Id: I91696e81e3cd8185729a388cd3b8bc45a298098f