videocodec CTS: Add support for running enc tests using fallback res
The existing resources used by videocodec cts apk are 1920x1080 level5
avc clips. If the decoder present on the device cannot decode this
clip, then most of the tests are skipped from running due to lack of
raw resource. Add a fallback resource, the tests can use in this case.

Bug: 290080485
Test: atest android.videocodec.cts

Change-Id: I9a282cd8dcfccddc634c3a6b1730c07215b662d8