HDMI: Handle KEYCODE_SETTINGS properly when Activity is sent to backround
The SETUP_MENU key might open the notification panel. This was
introduced in ag/24772642. We have to handle this key separately because
`onUserLeaveHint`, called when the activity is moved to background, was overriden for the Home key and there is no way to
capture what key triggered this event.
The alternative would be to have a common log for both these buttons in `onUserLeaveHint`,
but there is no clear way to differentiate them and it would introduce
more complexity if `config_shortPressOnSettingsBehavior` was set to 0.

Bug: 314713562
Test: run cts -m CtsHdmiCecHostTestCases
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:6bf33034354c7194e0c32b49e78389de0f96028c)
Merged-In: I0d8b223eccb8522b5d48ae7ae7febae3b8b9fe87
Change-Id: I0d8b223eccb8522b5d48ae7ae7febae3b8b9fe87