mediacuj CTS: Fix Seek test timeout in large test
In seek large test we are seeking forward 10 times, backward 10
times and then randomly seeking backward or forward 10 times.
While doing random seek we are using random function which can
leads to seek only backward all 10 times for which we didn't
calibarated our test timeout.
As we are having 3 clips for 5 min and 1 clip for 30 min, incresing
timeout as 300 sec and 100 sec respectively.

Bug: 316749293
Bug: 319390051

Test: atest CtsMediaCUJLargeTest
Change-Id: Ief598a3261b7fcfa0701116fbea5306fcf6d7eb7