CtsOsTestCases: Test fails due to Text change for Allow in permissions
To grant the permissions, the option with text 'Allow' is clicked in the test cases.
But when the test is changed to as per the customer reqirement the test case fails.
Fix is to find the text using text ID and click the option. Finding the option using text ID is already available.
Reused same to fix the issue.

Bug: 311318326
Test: Executed the CtsOsTestCases module test cases
Change-Id: Ie8fc422c162fb687aa9854ded1036fd4d21683b1
Merged-In: Ib604b27b681ce0b869f1532a2b2e3809fabce5b1