Remove channel mask matching from USB audio attributes test.
The USB device may only report supported channel counts but not channel
masks. In that case, it is up to the HAL to map channel count to channel
mask. Different HALs may have different mapping. In that case, testing
if the reported channel masks match to cached value may fail.

Bug: 315181258
Test: run the test
Change-Id: Ic236261f716ea3e9e2dee06479b4fe69a90384f2