Camera: Do not force STABILIZATION_MODE to OFF
Forcing STABILIZATION_MODE to OFF exposes crashes on some
HALs. Revert that to avoid the test regression.

Instead, skip zoom and crop check when stabilization is ON
in the capture template.

Keep stabilization as OFF for SETTINGS_OVERRIDE_ZOOM test
because without it we lost the ability to correlate zoom ratios.

Test: CtsCameraTestCases on Pixel factory and latest builds
Bug: 319391847
Change-Id: Id7aa8904e9992028248bf920d53b781da7047b32