Camera: Do not force STABILIZATION_MODE to OFF (take 2)
Forcing STABILIZATION_MODE to OFF exposes crashes on some
HALs. Revert that to avoid the test regression.

Instead, skip zoom and crop check when stabilization is ON
in the capture template.

Keep stabilization as OFF for SETTINGS_OVERRIDE_ZOOM test
because without it we lost the ability to correlate zoom ratios.

Compared to previous take, this change skips checking of cropRegion
monotonicity when stabilization is not OFF.

Test: CtsCameraTestCases on Pixel factory and latest builds
Bug: 319391847
Bug: 324508282
Change-Id: I51528bd44e924f947fab28a500de1a176019195f