Adjust currentMetrics logic per partially decoupled logic
On U+ devices, display cutout is not always considered a decor type of
the insets, and may not be excluded from the appBounds in
windowConfiguration, where the displayMetrics of the test come from.

This is controlled by a device config field. Treat not found as default
value to make the test compatible to existing U devices.

Bug: 321602048
Test: WindowMetricsWindowContextTests
Test: WindowMetricsActivityTests
Change-Id: Id8567035523ea585854825e9c4205591fa2332c3