CtsInputTestCases: Add skeleton for PointerIconTest
The PointerIconTest in CtsInputTestCases will serve as the end-to-end
test for the pointer icon pipeline. We will create a virtual display,
create and move a virtual mouse, set the pointer icon from an activity
on the display, and ensure that the icon is drawn correctly by comparing
it with the expected screenshot.

This CL adds the skeleton for the test, which does not include the
screenshot logic.

Bug: 324288919
Test: atest CtsInputTestCases:PointerIconTest
Change-Id: If1f697191f9ea8024af04d2d08e9379738260ea2