CTS tests for ComponentCaller#checkContentUriPermission in onNewIntent
These tests test the new Activity#getCurrentCaller and
Activity#onNewIntent(Intent, ComponentCaller), specifically for
ComponentCaller#checkContentUriPermission API.

The helper app and the test class each have 2 singleInstance activities
that are launched twice to ensure the #onNewIntent method is reached.
Once it's reached the test methods check the expected exceptions or
permission results.

To ensure the new intent caller is implemented correctly, when launching
the activity for the first time, the URI passed or the mode flags that
are sent are supposed to have the opposite effect on the actual result.

Bug: 293467489
Test: atest CtsAndroidAppTestCases:android.app.cts.ComponentCallerTest
Change-Id: I42991db59768ddd2f03a7d6d841e338ca575fadd