resourcesloader: add assets into assets attribute
This change adds assets into assets attribute of android_test. Before
the assets are added in a jar generated by a jar_genrule. After the
repackage change submitted, the jars will be run through jarjar, which
will compress the assets in the jar file. This change explicitly put the
assets in the assets attribute, and add aaptflags to avoid compressing
the assets file.

Test: atest CtsResourcesLoaderTests:android.content.res.loader.cts.ResourcesLoaderValuesTest#repeatedRemoveProviderNoOps
Bug: 326075332
Change-Id: Iaffe868c8a565fd5b8b3dc848e2c94f9180edfa4