Remove OrderedSelection CTS test
The test is actually testing 2 things at the same time:
1- The caller gets URIs for the images the user has clicked.
2- The caller gets URIs the user has clicked in order.

There's no reliable way to do 1) with the current design of the
photopicker, so the test tries to mimick that by creating files and then
pick the first images the photopicker presents. but the assumption
that the first images in the photopicker are the same as the images
created can be broken for many reasons including existing photos on the
device or a concurrent sync.

We are already testing 2) in SelectionTest.java, so removing this test
until photopicker redesign allows us to test both reliably.

Test: atest CtsPhotoPicker & atest com.android.providers.media.photopicker & manual tests
Bug: 324979956
Change-Id: I53e1c3f1fa6f9062e17d6201f7ae9cfc5ae85e70