Fix a bug for CarServiceHelperServiceTest.
CarServiceHelperServiceTest will timeout waiting for the new system
server to start if adb is in root because getIntProperty returns
a cached property value, not the latest value. So the property
sys.system_server.start_uptime remains the same even though system
server restarts.

In this CL, we use executeCommand instead so that there is no
caching.

Test: On seahawk, adb root; atest CarServiceHelperServiceTest
Bug: 326530175
Change-Id: Ie8163aef49eab866219cb9cf8dd2a587ad9f87ec