Fix CarRemoteAccessManagerTest.
Add require flag annotation for FLAG_CAR_DUMP_TO_PROTO since we
rely on that to check whether the client is a serverless client.

This CL also reorders setSelfAsServerlessClient and
assumeTaskSchedulingSupported because task scheduling is only
supported when the client is a serverless client.

Test: Follow CTS test plan to enable remote access feature on
emulator and run atest android.car.cts.CarRemoteAccessManagerTest
Bug: 325537480

Change-Id: Ieb88249dd2f8a6d0bf1a2846eba301c2cade03d5